Hazard Pot v1 / Hazard Projectile v1
- Created the hazard pot that will spawn a hazard tile when hit
- Also created a projectile prefab that when colliding with something, will create hazard tiles around it
