--- Assets/Scripts/Dungeons/Hazards/ProjectileThrow.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b4f64f4599893534aa4f5b860d01820c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
