--- Assets/Scripts/Dungeons/Hazards/TrajectoryPredictor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f937aa3f5f2f89340a33df4dd546fc94
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
